+++ demo/EFCoreDemo/EFCore/Repository/ITestEntityRepository.cs
@@ -1,11 +1,12 @@
-﻿using EasyCore.EFCoreRepository.IRepository;
-using EFCore.Entity;
 using EasyCore.Dependencie;
 using EasyCore.EFCoreRepository.DataFilter;
+using EasyCore.EFCoreRepository.IRepository;
+using EFCore.Entity;
+using EFCoreDbContext.EntityFrameworkCore.EFDbContext;
 
 namespace EFCore.Repository
 {
-    public interface ITestEntityRepository : IRepository<TestEntity>, ITransientDependencie
+    public interface ITestEntityRepository : IRepository<TestDbContext, TestEntity>, ITransientDependencie
     {
         List<IDataFilter> TestCustomEntityGetDataFilters();
     }